refactor(header): creation d'un comp Navbar
